third_party/blink/web_tests/external/wpt/html/browsers/the-window-object/window-prototype-chain-expected.txt

This is a testharness.js-based test.
PASS window object
FAIL Window.prototype assert_class_string: expected "[object WindowPrototype]" but got "[object Window]"
PASS Global scope polluter
FAIL EventTarget.prototype assert_class_string: expected "[object EventTargetPrototype]" but got "[object EventTarget]"
PASS Object.prototype
Harness: the test ran to completion.
